+++ transforms/mozclass.js
@@ -20,7 +20,7 @@ module.exports = (file, api, options) => {
 
   const printOptions =
     options.printOptions || {
-      quote: 'single',
+      quote: 'double',
       trailingComma: true,
       flowObjectCommas: true,
       arrowParensAlways: true,
